# Conflicts:
#	confignode/src/main/java/org/apache/iotdb/confignode/manager/node/NodeManager.java
#	server/src/main/java/org/apache/iotdb/db/service/DataNode.java

+++ confignode/src/main/java/org/apache/iotdb/confignode/persistence/partition/RegionGroup.java
@@ -78,7 +78,7 @@ public class RegionGroup {
         }));
   }
 
-  public long getSeriesSlotCount() {
+  public int getSeriesSlotCount() {
     return slotCountMap.size();
   }
